Refactor to job instead of notification handler
