+++ src/Umbraco.Cms.Search.Core/Services/ContentIndexing/IContentIndexingService.cs
@@ -20,6 +20,5 @@ public interface IContentIndexingService
     /// </summary>
     /// <param name="indexAlias">The alias of the index to rebuild.</param>
     /// <param name="origin">An identifier for the server/request that requested the rebuild, used for same-origin filtering.</param>
-    /// <param name="delay">An optional delay to wait before the queued rebuild actually starts work.</param>
-    void Rebuild(string indexAlias, string origin, TimeSpan? delay = null);
+    void Rebuild(string indexAlias, string origin);
 }
