+++ src/Umbraco.Cms.Search.Provider.Examine/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -11,13 +11,14 @@ using Umbraco.Cms.Api.Management.OpenApi;
 using Umbraco.Cms.Core.Configuration.Models;
 using Umbraco.Cms.Core.DependencyInjection;
 using Umbraco.Cms.Core.Hosting;
-using Umbraco.Cms.Core.Notifications;
 using Umbraco.Cms.Infrastructure.Telemetry.Interfaces;
 using Umbraco.Cms.Search.Core.Notifications;
+using Umbraco.Cms.Search.Provider.Examine.BackgroundJobs;
 using Umbraco.Cms.Search.Provider.Examine.Lucene;
 using Umbraco.Cms.Search.Provider.Examine.NotificationHandlers;
 using Umbraco.Cms.Search.Provider.Examine.Services;
 using Umbraco.Cms.Search.Provider.Examine.Telemetry;
+using Umbraco.Extensions;
 
 namespace Umbraco.Cms.Search.Provider.Examine.DependencyInjection;
 
@@ -92,7 +93,7 @@ public static class UmbracoBuilderExtensions
                     s, new DirectoryInfo(tempDir), s.GetRequiredService<IApplicationRoot>().ApplicationRoot);
             });
 
-        builder.AddNotificationHandler<UmbracoApplicationStartedNotification, RebuildNotificationHandler>();
+        builder.Services.AddRecurringBackgroundJob<IndexRebuildJob>();
 
         builder.Services.AddTransient<IDetailedTelemetryProvider, ExamineTelemetryProvider>();
